requirements.txt
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
djoser==2.0.3
django-widget-tweaks~=1.4.8